Reorganize watchers into service_watcher/type/type
