+++ lib/synapse/service_watcher/base/base.rb
