+++ lib/synapse/service_watcher/docker/docker.rb
@@ -1,4 +1,4 @@
-require "synapse/service_watcher/base"
+require "synapse/service_watcher/base/base"
 require 'docker'
 
 class Synapse::ServiceWatcher
