+++ lib/synapse/service_watcher/ec2tag/ec2tag.rb
@@ -1,4 +1,4 @@
-require 'synapse/service_watcher/base'
+require 'synapse/service_watcher/base/base'
 require 'aws-sdk'
 
 class Synapse::ServiceWatcher
