+++ lib/synapse/service_watcher/multi/multi.rb
@@ -1,4 +1,4 @@
-require "synapse/service_watcher/base"
+require "synapse/service_watcher/base/base"
 require "synapse/log"
 require "synapse/statsd"
