+++ lib/synapse/service_watcher/zookeeper/zookeeper.rb
@@ -1,4 +1,4 @@
-require "synapse/service_watcher/base"
+require "synapse/service_watcher/base/base"
 
 require 'thread'
 require 'zk'
